\name{FindFormat}
\alias{FindFormat}

\title{
	Find the format the date is written in 
}
\description{
	An internal function that finds the R-code formats for dates. 
}
\usage{
FindFormat(string, CurrentDate)
}

\arguments{
  \item{string}{
	The current date in a format chosen through the GUI.
}
  \item{CurrentDate}{
	A vector containing the current date in a variety of formats. 
}
}


\value{

 \item{string}{The R-code for the format of the date.}


}

\author{
	Joe Swintek
}

