\name{generateExamples}
\alias{generateExamples}

\title{
	Generate examples
}
\description{
	Creates a folder which is populated with examples of appropriate data sets and their corresponding analyses for every type of analysis StatCharrms is capable of producing.      
}
\usage{
generateExamples()
}

\details{
	Called by pressing the [\bold{Generate Examples}] button from the main StatCharrms window.   
}

\author{
	Joe Swintek
}
\note{
	This will create and populate a folder on the computers hard drive. 
}

