\name{getFileName}
\alias{getFileName}

\title{
	Gets a file name
}
\description{
	Calls a dialogue window where the name of a file to be saved can be typed in.  
}
\usage{
getFileName()
}

\details{
	This is called immediately before a data structure is saved.
}
\value{

	\item{FileName}{The name written into the dialogue window.}


}


\author{
	Joe Swintek
}

