\name{ReadMeFile}
\alias{ReadMeFile}

\title{
	Calls the read me file
}
\description{
	Call a file containing the authors, citation, and changes of StatCharrms.
}
\usage{
ReadMeFile()
}

\value{
 \item{\bold{Changes button}		}{ Calls \code{\link{ShowChangeLog}} which displays the changes between versions.}
 \item{\bold{References button}}{Calls \code{\link{Reference}} which displays the references for each package StatCharrms uses.}
}

\author{
	Joe Swintek
}
