\name{Reference}
\alias{Reference}

\title{
	Display the references for StatCharrms
}
\description{
	This is used to display the citations for all the packages StatCharrms uses.
}
\usage{
Reference()
}

\details{
	Called by the [\bold{References}] button in the \code{\link{ReadMeFile}} window.
}

\author{
	Joe Swintek
}