\name{ShowChangeLog}
\alias{ShowChangeLog}

\title{
	Display the change logs for StatCharms
}
\description{
	Displays the changes that have occurred between the versions of StatCharrms.  Called from the [\bold{changes}] button 
	from the "Read me" window.
}
\usage{
ShowChangeLog()
}

\author{
	Joe Swintek
}

\seealso{
	Also see  \code{\link{ReadMeFile}}
}
