\name{StatCharrms-package}
\alias{StatCharrms-package}
\alias{StatCharrms}
\docType{package}
\title{
	StatCharrms
}
\description{
   Statistical analysis of Chemistry, Histopathology, and Reproduction endpoints including Repeated measures and Multi-generation Studies (StatCharrms).
	Based on the SAS build of StatCharrms written by John Green, the R version of 
	StatCharrms was developed for the statistical analyses needed for the tier II endocrine disruptor screening program. StatCharrms was funded by the USEPA under Contract EP-D-13-052.  
}

\details{
\itemize{
\item{\bold{Package:}}{ StatCharrms}
\item{\bold{Type:}}{Package \ Frontend}
\item{\bold{Version:}}{0.96-00}
\item{\bold{Date:}}{   2016-06-20}
\item{\bold{License:}}{   CC0 }
}
}

\author{
Joe Swintek <swintek.joe@epa.gov> as author, Kevin Flynn as a tester, Jon Haselman as a tester. 

Maintainer: Joe Swintek <swintek.joe@epa.gov>
}

\references{
OECD Guidelines for the Testing of Chemicals, Section 2. Test No. 240: Medaka Extended One Generation Reproduction Test (MEOGRT)
}

\keyword{package}

\examples{
\dontrun{
Run.StatCharrms()
generateExamples()
}
}


